+++ Tests/MapboxSearchTests/Common/Data Samples/BoundingBox+Samples.swift
@@ -2,4 +2,5 @@
 
 extension BoundingBox {
     static let sample1 = BoundingBox(.sample1, .sample2)
+    static let sample2 = BoundingBox(.sample2, .sample3)
 }